requirements.txt
# Dev environment
pip
autopep8
python-dotenv

# App
flask
flask-sqlalchemy
flask-login
werkzeug
Pillow
opencv-python
email-validator

# Database (PostgreSQL)
psycopg2-binary

# Email with Flask
Flask-Mail